money to float et debut AbsParser

+++ src/main/kotlin/poklin/Bet.kt
@@ -2,4 +2,4 @@ package poklin
 
 import poklin.model.bet.BettingDecision.BettingAction
 
-class Bet(var action: BettingAction, var amount: Int = 0)
+class Bet(var action: BettingAction, var amount: Float = 0f)
